// apps/frontend/src/gen/vibesync/common/v1/common_pb.ts
export enum RoomRole {
  /**
   * @generated from enum value: ROOM_ROLE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: ROOM_ROLE_GUEST = 1;
   */
  GUEST = 1,

  /**
   * @generated from enum value: ROOM_ROLE_MEMBER = 2;
   */
  MEMBER = 2,

  /**
   * @generated from enum value: ROOM_ROLE_MODERATOR = 3;
   */
  MODERATOR = 3,

  /**
   * @generated from enum value: ROOM_ROLE_OWNER = 4;
   */
  OWNER = 4,
}